Got important database stuff working and moved database-related files into a package
