+++ app/build.gradle.kts
@@ -59,7 +59,8 @@ dependencies {
 	implementation(libs.androidx.lifecycle.viewmodel.compose)
 	implementation(libs.androidx.room.runtime)
 	implementation(libs.androidx.room.ktx)
-	ksp(libs.androidx.room.compiler)
+    implementation(libs.runtime.livedata)
+    ksp(libs.androidx.room.compiler)
 	implementation(libs.androidx.core.ktx)
 	implementation(libs.androidx.lifecycle.runtime.ktx)
 	implementation(libs.androidx.activity.compose)
