+++ app/src/main/java/com/example/csce546_project/MainScreen.kt
@@ -21,6 +21,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
@@ -30,7 +31,6 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -39,6 +39,7 @@ import androidx.compose.ui.window.Popup
 import androidx.core.content.ContextCompat
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.viewmodel.compose.viewModel
+import com.example.csce546_project.database.PictureEntry
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.isGranted
 import com.google.accompanist.permissions.rememberPermissionState
@@ -205,22 +206,22 @@ fun MainScreen() {
 			Column(
 				modifier = Modifier.fillMaxWidth().fillMaxHeight().background(color = Color.Red)
 			) {
-				Text(
-					text = "TODO picture list",
-					fontWeight = FontWeight.Bold,
-					fontSize = 36.sp
-				)
-
 				Button(
 					onClick = { viewModel.openAddPopup() },
 				) {
 					Text("Test add popup")
 				}
 
-				Button(
-					onClick = { /* TODO open edit popup */ },
-				) {
-					Text("Test view popup")
+				val pictures by viewModel.pictures.observeAsState(emptyList())
+
+				pictures.forEach { picture ->
+					Button(
+						onClick = { viewModel.openEditPopup(picture) }
+					) {
+						Text(
+							text = ("ID=" + picture.id + " | NAME=" + (picture.name ?: "NULL"))
+						)
+					}
 				}
 			}
 		}
